Move heartbeat into the main class
