+++ ts/index.ts
@@ -2,7 +2,6 @@ import Pubsub, { InMessage, utils } from 'libp2p-interfaces/src/pubsub'
 import { MessageCache } from './message-cache'
 import { RPC, IRPC } from './message/rpc'
 import * as constants from './constants'
-import { Heartbeat } from './heartbeat'
 import { getGossipPeers } from './get-gossip-peers'
 import { createGossipRpc, shuffle, hasGossipProtocol, messageIdToString } from './utils'
 import {
@@ -196,9 +195,6 @@ class Gossipsub extends Pubsub {
   /** Peer score tracking */
   score: PeerScore
 
-  /** A heartbeat timer that maintains the mesh */
-  heartbeat: Heartbeat
-
   /**
    * Number of heartbeats since the beginning of time
    * This allows us to amortize some resource cleanup -- eg: backoff cleanup
@@ -222,6 +218,12 @@ class Gossipsub extends Pubsub {
   emit: (event: string | symbol, ...args: any[]) => boolean
 
   public static multicodec: string = constants.GossipsubIDv11
+
+  _heartbeatTimer: {
+    _intervalId: NodeJS.Timeout | undefined
+    runPeriodically(fn: () => void, period: number): void
+    cancel(): void
+  } | null
 
   // TODO: add remaining props
   /**
@@ -293,8 +295,6 @@ class Gossipsub extends Pubsub {
     this.getFastMsgIdStr = options.fastMsgIdFn ?? undefined
 
     this.fastMsgIdCache = options.fastMsgIdFn ? new SimpleTimeCache<string>({ validityMs: opts.seenTTL }) : undefined
-
-    this.heartbeat = new Heartbeat(this)
 
     /**
      * libp2p
@@ -892,7 +892,6 @@ class Gossipsub extends Pubsub {
    */
   async start(): Promise<void> {
     await super.start()
-    this.heartbeat.start()
     this.score.start()
     // connect to direct peers
     this._directPeerInitial = setTimeout(() => {
@@ -900,6 +899,26 @@ class Gossipsub extends Pubsub {
         this._connect(id)
       })
     }, constants.GossipsubDirectConnectInitialDelay)
+
+    if (!this._heartbeatTimer) {
+      const heartbeat = this._heartbeat.bind(this)
+
+      const timeout = setTimeout(() => {
+        heartbeat()
+        this._heartbeatTimer!.runPeriodically(heartbeat, this._options.heartbeatInterval)
+      }, constants.GossipsubHeartbeatInitialDelay)
+
+      this._heartbeatTimer = {
+        _intervalId: undefined,
+        runPeriodically: (fn, period) => {
+          this._heartbeatTimer!._intervalId = setInterval(fn, period)
+        },
+        cancel: () => {
+          clearTimeout(timeout)
+          clearInterval(this._heartbeatTimer!._intervalId as NodeJS.Timeout)
+        },
+      }
+    }
   }
 
   /**
@@ -908,7 +927,12 @@ class Gossipsub extends Pubsub {
    */
   async stop(): Promise<void> {
     await super.stop()
-    this.heartbeat.stop()
+
+    if (this._heartbeatTimer) {
+      this._heartbeatTimer.cancel()
+      this._heartbeatTimer = null
+    }
+
     this.score.stop()
 
     this.mesh = new Map()
@@ -1384,6 +1408,266 @@ class Gossipsub extends Pubsub {
       peers: px,
       backoff: backoff,
     }
+  }
+
+  /**
+   * Maintains the mesh and fanout maps in gossipsub.
+   */
+  _heartbeat(): void {
+    const { D, Dlo, Dhi, Dscore, Dout, fanoutTTL } = this._options
+    this.heartbeatTicks++
+
+    // cache scores throught the heartbeat
+    const scores = new Map<string, number>()
+    const getScore = (id: string): number => {
+      let s = scores.get(id)
+      if (s === undefined) {
+        s = this.score.score(id)
+        scores.set(id, s)
+      }
+      return s
+    }
+
+    // peer id => topic[]
+    const tograft = new Map<string, string[]>()
+    // peer id => topic[]
+    const toprune = new Map<string, string[]>()
+    // peer id => don't px
+    const noPX = new Map<string, boolean>()
+
+    // clean up expired backoffs
+    this._clearBackoff()
+
+    // clean up peerhave/iasked counters
+    this.peerhave.clear()
+    this.iasked.clear()
+
+    // apply IWANT request penalties
+    this._applyIwantPenalties()
+
+    // ensure direct peers are connected
+    this._directConnect()
+
+    // maintain the mesh for topics we have joined
+    this.mesh.forEach((peers, topic) => {
+      // prune/graft helper functions (defined per topic)
+      const prunePeer = (id: string): void => {
+        this.log('HEARTBEAT: Remove mesh link to %s in %s', id, topic)
+        // update peer score
+        this.score.prune(id, topic)
+        // add prune backoff record
+        this._addBackoff(id, topic)
+        // remove peer from mesh
+        peers.delete(id)
+        // add to toprune
+        const topics = toprune.get(id)
+        if (!topics) {
+          toprune.set(id, [topic])
+        } else {
+          topics.push(topic)
+        }
+      }
+      const graftPeer = (id: string): void => {
+        this.log('HEARTBEAT: Add mesh link to %s in %s', id, topic)
+        // update peer score
+        this.score.graft(id, topic)
+        // add peer to mesh
+        peers.add(id)
+        // add to tograft
+        const topics = tograft.get(id)
+        if (!topics) {
+          tograft.set(id, [topic])
+        } else {
+          topics.push(topic)
+        }
+      }
+
+      // drop all peers with negative score, without PX
+      peers.forEach((id) => {
+        const score = getScore(id)
+        if (score < 0) {
+          this.log('HEARTBEAT: Prune peer %s with negative score: score=%d, topic=%s', id, score, topic)
+          prunePeer(id)
+          noPX.set(id, true)
+        }
+      })
+
+      // do we have enough peers?
+      if (peers.size < Dlo) {
+        const backoff = this.backoff.get(topic)
+        const ineed = D - peers.size
+        const peersSet = getGossipPeers(this, topic, ineed, (id) => {
+          // filter out mesh peers, direct peers, peers we are backing off, peers with negative score
+          return !peers.has(id) && !this.direct.has(id) && (!backoff || !backoff.has(id)) && getScore(id) >= 0
+        })
+
+        peersSet.forEach(graftPeer)
+      }
+
+      // do we have to many peers?
+      if (peers.size > Dhi) {
+        let peersArray = Array.from(peers)
+        // sort by score
+        peersArray.sort((a, b) => getScore(b) - getScore(a))
+        // We keep the first D_score peers by score and the remaining up to D randomly
+        // under the constraint that we keep D_out peers in the mesh (if we have that many)
+        peersArray = peersArray.slice(0, Dscore).concat(shuffle(peersArray.slice(Dscore)))
+
+        // count the outbound peers we are keeping
+        let outbound = 0
+        peersArray.slice(0, D).forEach((p) => {
+          if (this.outbound.get(p)) {
+            outbound++
+          }
+        })
+
+        // if it's less than D_out, bubble up some outbound peers from the random selection
+        if (outbound < Dout) {
+          const rotate = (i: number): void => {
+            // rotate the peersArray to the right and put the ith peer in the front
+            const p = peersArray[i]
+            for (let j = i; j > 0; j--) {
+              peersArray[j] = peersArray[j - 1]
+            }
+            peersArray[0] = p
+          }
+
+          // first bubble up all outbound peers already in the selection to the front
+          if (outbound > 0) {
+            let ihave = outbound
+            for (let i = 1; i < D && ihave > 0; i++) {
+              if (this.outbound.get(peersArray[i])) {
+                rotate(i)
+                ihave--
+              }
+            }
+          }
+
+          // now bubble up enough outbound peers outside the selection to the front
+          let ineed = D - outbound
+          for (let i = D; i < peersArray.length && ineed > 0; i++) {
+            if (this.outbound.get(peersArray[i])) {
+              rotate(i)
+              ineed--
+            }
+          }
+        }
+
+        // prune the excess peers
+        peersArray.slice(D).forEach(prunePeer)
+      }
+
+      // do we have enough outbound peers?
+      if (peers.size >= Dlo) {
+        // count the outbound peers we have
+        let outbound = 0
+        peers.forEach((p) => {
+          if (this.outbound.get(p)) {
+            outbound++
+          }
+        })
+
+        // if it's less than D_out, select some peers with outbound connections and graft them
+        if (outbound < Dout) {
+          const ineed = Dout - outbound
+          const backoff = this.backoff.get(topic)
+          getGossipPeers(this, topic, ineed, (id: string): boolean => {
+            // filter our current mesh peers, direct peers, peers we are backing off, peers with negative score
+            return !peers.has(id) && !this.direct.has(id) && (!backoff || !backoff.has(id)) && getScore(id) >= 0
+          }).forEach(graftPeer)
+        }
+      }
+
+      // should we try to improve the mesh with opportunistic grafting?
+      if (this.heartbeatTicks % constants.GossipsubOpportunisticGraftTicks === 0 && peers.size > 1) {
+        // Opportunistic grafting works as follows: we check the median score of peers in the
+        // mesh; if this score is below the opportunisticGraftThreshold, we select a few peers at
+        // random with score over the median.
+        // The intention is to (slowly) improve an underperforming mesh by introducing good
+        // scoring peers that may have been gossiping at us. This allows us to get out of sticky
+        // situations where we are stuck with poor peers and also recover from churn of good peers.
+
+        // now compute the median peer score in the mesh
+        const peersList = Array.from(peers).sort((a, b) => getScore(a) - getScore(b))
+        const medianIndex = Math.floor(peers.size / 2)
+        const medianScore = getScore(peersList[medianIndex])
+
+        // if the median score is below the threshold, select a better peer (if any) and GRAFT
+        if (medianScore < this._options.scoreThresholds.opportunisticGraftThreshold) {
+          const backoff = this.backoff.get(topic)
+          const peersToGraft = getGossipPeers(
+            this,
+            topic,
+            constants.GossipsubOpportunisticGraftPeers,
+            (id: string): boolean => {
+              // filter out current mesh peers, direct peers, peers we are backing off, peers below or at threshold
+              return (
+                peers.has(id) && !this.direct.has(id) && (!backoff || !backoff.has(id)) && getScore(id) > medianScore
+              )
+            }
+          )
+          peersToGraft.forEach((id) => {
+            this.log('HEARTBEAT: Opportunistically graft peer %s on topic %s', id, topic)
+            graftPeer(id)
+          })
+        }
+      }
+
+      // 2nd arg are mesh peers excluded from gossip. We have already pushed
+      // messages to them, so its redundant to gossip IHAVEs.
+      this._emitGossip(topic, peers)
+    })
+
+    // expire fanout for topics we haven't published to in a while
+    const now = this._now()
+    this.lastpub.forEach((lastpb, topic) => {
+      if (lastpb + fanoutTTL < now) {
+        this.fanout.delete(topic)
+        this.lastpub.delete(topic)
+      }
+    })
+
+    // maintain our fanout for topics we are publishing but we have not joined
+    this.fanout.forEach((fanoutPeers, topic) => {
+      // checks whether our peers are still in the topic and have a score above the publish threshold
+      const topicPeers = this.topics.get(topic)
+      fanoutPeers.forEach((id) => {
+        if (!topicPeers!.has(id) || getScore(id) < this._options.scoreThresholds.publishThreshold) {
+          fanoutPeers.delete(id)
+        }
+      })
+
+      // do we need more peers?
+      if (fanoutPeers.size < D) {
+        const ineed = D - fanoutPeers.size
+        const peersSet = getGossipPeers(this, topic, ineed, (id: string): boolean => {
+          // filter out existing fanout peers, direct peers, and peers with score above the publish threshold
+          return (
+            !fanoutPeers.has(id) &&
+            !this.direct.has(id) &&
+            getScore(id) >= this._options.scoreThresholds.publishThreshold
+          )
+        })
+        peersSet.forEach((id) => {
+          fanoutPeers.add(id)
+        })
+      }
+
+      // 2nd arg are fanout peers excluded from gossip.
+      // We have already pushed messages to them, so its redundant to gossip IHAVEs
+      this._emitGossip(topic, fanoutPeers)
+    })
+
+    // send coalesced GRAFT/PRUNE messages (will piggyback gossip)
+    this._sendGraftPrune(tograft, toprune, noPX)
+
+    // flush pending gossip that wasn't piggybacked above
+    this._flush()
+
+    // advance the message history window
+    this.messageCache.shift()
+
+    this.emit('gossipsub:heartbeat')
   }
 }
